Додана можливість додавати акторів в базу данних через інтерфейс веб-сторінки.

+++ Cinema/BusinessLogic/DTOs/ActorDTO.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string Surname { get; set; }
-        public ushort BirthYear { get; set; }
+        public int BirthYear { get; set; }
         public string PhotoUrl { get; set; }
         public List<RoleDTO> Roles { get; set; }
     }
